minor changes to styling
