+++ src/components/CharacterSheet.js
@@ -5,6 +5,8 @@ import {Row, Col, Panel} from 'react-bootstrap';
 
 import {SKILLS} from '../../lib/Character';
 
+import '../stylesheets/components/CharacterSheet';
+
 const ICONS = {
   savingThrows: {
     true: 'star',
@@ -27,39 +29,6 @@ function getIcon(name, data) {
   }
 }
 
-export class AbilityScores extends React.Component {
-  constructor(props) {
-    super(props);
-  }
-
-  render() {
-    var abilityScores = [];
-    for (let abilityScore in this.props.abilityScores) {
-      abilityScores.push(
-        <AbilityScore
-          key={abilityScore}
-          name={abilityScore}
-          savingThrows={this.props.savingThrows[abilityScore]}
-          skills={this.props.skills}
-          value={this.props.abilityScores[abilityScore]}
-        />
-      );
-    }
-
-    return (
-      <div className="ability-scores">
-        {abilityScores}
-      </div>
-    );
-  }
-}
-
-AbilityScores.propTypes = {
-  abilityScores: React.PropTypes.object.isRequired,
-  savingThrows: React.PropTypes.object.isRequired,
-  skills: React.PropTypes.object.isRequired
-};
-
 export class AbilityScore extends React.Component {
 	constructor(props) {
 		super(props);
@@ -80,7 +49,12 @@ export class AbilityScore extends React.Component {
       <Row className="ability-score" id={name} >
         <Col className="col" md={4}>
           <Panel header={name} className="centered">
-            {this.props.value}
+            <div className='ability-score value'>
+              {this.props.value}
+            </div>
+            <div className='ability-score mod'>
+              {this.props.mod}
+            </div>
           </Panel>
         </Col>
         <Col className="col" md={8}>
@@ -99,6 +73,7 @@ export class AbilityScore extends React.Component {
 }
 
 AbilityScore.propTypes = {
+  mod: React.PropTypes.number.isRequired,
 	name: React.PropTypes.string.isRequired,
   savingThrows: React.PropTypes.object.isRequired,
   skills: React.PropTypes.object.isRequired,
