+++ src/stylesheets/containers/CharacterSheet.scss
@@ -3,19 +3,6 @@
   width: 100%;
 }
 
-.panel {
-  .fa {
-    margin-right: 5px;
-  }
-}
-
-.panel.centered {
-  .panel-heading, .panel-body {
-    text-align: center;
-    padding: 10px 0;
-  }
-}
-
 .character-sheet-grid {
   .col {
     background: #ffffff;
